requirements.txt
mysql-connector-python==9.0.0
pymongo==4.8.0
